Move to GroundItemsPlugin
