--- runelite-client/src/main/java/net/runelite/client/plugins/groupitemlist/GroupItemListConfig.java
@@ -1,85 +1,0 @@
-/*
- * Copyright (c) 2019, Ricky <https://github.com/RickyLoader>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package net.runelite.client.plugins.groupitemlist;
-
-import net.runelite.client.config.Config;
-import net.runelite.client.config.ConfigGroup;
-import net.runelite.client.config.ConfigItem;
-
-@ConfigGroup("groupitemlist")
-public interface GroupItemListConfig extends Config
-{
-
-	enum positionOption
-	{
-		PREPEND,
-		APPEND
-	}
-
-	@ConfigItem(
-		keyName = "quantityPosition",
-		name = "Quantity position",
-		description = "Configures where to place the item's quantity",
-		position = 0
-	)
-
-	default positionOption getQuantityPosition()
-	{
-		return positionOption.APPEND;
-	}
-
-	enum styleOption
-	{
-		SQUARE,
-		PARENTHESES,
-		CURLY,
-		NONE
-	}
-
-	@ConfigItem(
-		keyName = "quantityStyle",
-		name = "Quantity style",
-		description = "Configures how the quantity will be styled",
-		position = 1
-	)
-
-	default styleOption getQuantityStyle()
-	{
-		return styleOption.SQUARE;
-	}
-
-	@ConfigItem(
-		keyName = "includeX",
-		name = "Include x",
-		description = "Configures whether to include an x before the quantity: 'Shark x32'",
-		position = 2
-	)
-
-	default boolean includeX()
-	{
-		return false;
-	}
-
-}
